Add redux toolkit and styled-componentes, Move page files into their index.js instead of exporting
