+++ src/components/searchBar/index.js
