+++ src/pages/movieDetails/index.js
@@ -1,1 +1,51 @@
-export { default } from './movieDetails';
+import React, { useEffect, useState } from 'react'
+import { Link } from 'react-router-dom'
+import { apiKey } from '../../constants'
+import './movieDetails.css'
+
+const MovieDetails = ({match}) => {
+    const movie = useMovie(match)
+
+	return(		
+        <div className="bg-white p-4 rounded row movie-details">
+            <div className="col-xs-12 col-md-4">
+                <img className="img-fluid" src={"https://image.tmdb.org/t/p/w300/" + movie.poster_path} alt="" />						
+            </div>
+            <div className="col-xs-12 col-md-8" style={{ position: "relative" }}>
+                <p><b>Overview:</b> {movie.overview}</p>
+                <p><b>Release date:</b> {movie.release_date}</p>
+                <p><b>Original language:</b> {movie.original_language}</p>
+                <div className="row bg-dark text-white bottom-align-text">
+                    <div className="bg-success col-xs-3 col-md-3">
+                        <h1>{movie.vote_average}</h1>
+                        <p>Rating</p>
+                    </div>
+                    <div className="col-xs-9 col-md-9 p-2">
+                        <p><b>Vote count:</b> {movie.vote_count}</p>
+                        <Link className="btn btn-warning float-right" to="/movies">Back</Link>
+                    </div>
+                </div>
+            </div>
+        </div>
+	);
+}
+
+const useMovie = match => {
+    const [movie, setMovie] = useState("")
+
+    useEffect(() => {
+        fetchMovie()
+    }, [])
+    
+    const fetchMovie = async () => {
+        const link = 
+            `https://api.themoviedb.org/3/movie/${match.params.id}?api_key=${apiKey}&language=en-US` 
+        const data = await fetch(link)
+        const response = await data.json()
+        setMovie(response)
+    }
+
+    return movie
+}
+
+export default MovieDetails;
